requirements.txt
# Telegram Bot
aiogram==3.17.0

# Google API
google-auth==2.37.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
google-api-python-client==2.158.0

# Utils
python-dotenv==1.0.1
pytz==2024.2